requirements.txt
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cmake==3.27.7
colorama==0.4.6
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
google-api-core==2.13.0
google-auth==2.23.4
google-cloud-vision==3.4.5
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
idna==3.4
imageio==2.31.6
imutils==0.5.4
lazy_loader==0.3
networkx==3.1
numpy==1.24.4
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyWavelets==1.4.1
requests==2.31.0
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
tifffile==2023.7.10
urllib3==2.0.7